not use BigInt

+++ src/main/g8/core/src/main/scala/cs320/Value.scala
@@ -6,7 +6,7 @@ object Value {
 
   type Env = Map[String, Value]
 
-  case class IntV(value: Int) extends Value
+  case class IntV(value: BigInt) extends Value
   case class BooleanV(value: Boolean) extends Value
   case class TupleV(values: List[Value]) extends Value
   case object NilV extends Value
